LayoutTests/imported/w3c/web-platform-tests/eventsource/request-credentials.any.window-expected.txt

Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=1_1644427824158
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=1_1644427824158 due to access control checks.
Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=2_1644427824163
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=2_1644427824163 due to access control checks.
Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=3_1644427824165
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors-cookie.py?ident=3_1644427824165 due to access control checks.

Harness Error (TIMEOUT), message = null

TIMEOUT EventSource: credentials: credentials enabled Test timed out
TIMEOUT EventSource: credentials: credentials disabled Test timed out
TIMEOUT EventSource: credentials: credentials default Test timed out
